refactor: implement proper API router with controller pattern
- Replace large if/else chain in HTTPAPIService with clean router system
- Extract business logic into dedicated controllers (Tasks, Pomodoro, Time, Webhooks, System)
- Reduce HTTPAPIService from 1,061 to 231 lines (~78% reduction)
- Implement APIRouter class with parameter extraction for dynamic routes
- Adopt controller/service pattern for better maintainability and testability
- Maintain full backward compatibility with existing API endpoints
- All tests passing and functionality preserved

This architectural improvement makes adding new endpoints trivial and follows
modern web framework patterns for better code organization.
